src/main/java/eu/erasmuswithoutpaper/organization/entity/Institution.java

package eu.erasmuswithoutpaper.organization.entity;

import java.io.Serializable;
import java.util.List;
import javax.persistence.CascadeType;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.JoinColumn;
import javax.persistence.OneToMany;

@Entity
public class Institution implements Serializable{

    @Id
    private String institutionId;
    private String otherId;
    private String name;
    private String country;
    private String description;
    
    @OneToMany(cascade = CascadeType.ALL, orphanRemoval=true)
    @JoinColumn(name = "organization_unit_id")
    private List<OrganizationUnit> organinazationUnits;

    public String getInstitutionId() {
        return institutionId;
    }

    public void setInstitutionId(String institutionId) {
        this.institutionId = institutionId;
    }

    public String getOtherId() {
        return otherId;
    }

    public void setOtherId(String otherId) {
        this.otherId = otherId;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getCountry() {
        return country;
    }

    public void setCountry(String country) {
        this.country = country;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public List<OrganizationUnit> getOrganinazationUnits() {
        return organinazationUnits;
    }

    public void setOrganinazationUnits(List<OrganizationUnit> organinazationUnits) {
        this.organinazationUnits = organinazationUnits;
    }
    
}
